+++ src/powerquery-parser/parser/parsers/index.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import * as NaiveParseSteps from "./naive";
+import * as NaiveParseSteps from "./naiveParseSteps";
 
 export { NaiveParseSteps };
 export { CombinatorialParser } from "./combinatorialParser";
